cookies validated.. but used to set the state. needs improving > #46
