+++ client/src/helpers/handleLogout.js
@@ -1,0 +1,6 @@
+import Cookies from "js-cookie";
+function handleLogout() {
+  Cookies.remove("user_auth");
+}
+
+export default handleLogout;
