+++ client/src/helpers/validateCookie.js
@@ -1,27 +1,39 @@
-// const validateCookie = () => {
-//   var getCookiePromise = new Promise(function(resolve, reject) {
-//     const cookieStr = Cookies.get("user_auth");
-//     const cookieObj = { user_auth: cookieStr };
-//     resolve(cookieObj);
-//   });
-//   getCookiePromise
-//     .then(cookieObj => {
-//       fetch("/api/cookieValidation", {
-//         method: "POST",
-//         headers: {
-//           "Content-Type": "application/json; charset=utf-8"
-//         },
-//         body: JSON.stringify(cookieObj)
-//       })
-//         .then(res => res.json())
-//         .then(res => {
-//           console.log("IM IN APP.JS>>" + res.auth + "   " + res.username);
-//         })
-//         .catch(err => {
-//           console.log(err);
-//         });
-//     })
-//     .catch(err => console.log(err));
-// };
+const decodeCookie = new Promise(function(resolve, reject) {
+  fetch("/api/cookieValidation", {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json; charset=utf-8"
+    }
+  })
+    .then(res => res.json())
+    .then(res => {
+      resolve(res.username);
+    })
+    .catch(err => {
+      reject(err);
+    });
+});
 
-// module.export = { validateCookie };
+function validateCookieTF(that) {
+  decodeCookie
+    .then(function(decodedCookie) {
+      if (typeof decodedCookie === "string") {
+        that.setState({
+          auth: true,
+          authUser: decodedCookie
+        });
+      } else {
+        that.setState({
+          auth: false,
+          authUser: null
+        });
+      }
+    })
+    .catch(err => console.log(err));
+  return;
+}
+const validateCookieUser = decodeCookie
+  .then(decodedCookie => decodedCookie)
+  .catch(err => console.log(err));
+
+export default validateCookieTF;
